Fix name of zone created with wrapper
It was called "wrappedZone", due to a bug (js strings are false when
zero-length). I fixed this by implementing a name option, so zones can
be named explicitly, which is probably useful when you don't control the
name of the body function, and allowed me to simplify the places where
name decisions were being made, and get rid of the _zoneName property on
the body function.