zone: don't dump stack twice on invoke errors
Error was being printed as an object. It seems Error objects don't have
a `stack` property until it's referenced, at which point they do have
the property, and it gets dumped as a string property. Effect is that
	console.log('error:', e)
prints just the message string with some object annotation, and
	console.log('error:', e, e.stack}
prints an object with a very long string `.stack` property, and then the
stack.