require: don't load the `buffer` module twice
The buffer module contains nothing in need of monkeypatching nor does
it depend on anything that does.

At the same time the `Buffer` constructor is a global object that gets
loaded very early in node's startup process. If we allow the `buffer`
module to be loaded again it will create a new constructor, which breaks
instanceof checks and causes other unexpected behavior.